requirements.txt
pandas==2.3.2
numpy==2.2.6
google-generativeai==0.8.6
streamlit==1.52.2
matplotlib==3.10.6
python-dotenv==1.1.1
crewai==0.177.0
seaborn
scikit-learn
matplotlib